requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
qrcode==7.3.1
urllib3==1.26.10
yarl==1.7.2
pysondb==1.6.7
